Split skills subcommands into different files
