+++ src/main/java/com/archyx/aureliumskills/AureliumSkills.java
@@ -4,10 +4,7 @@ import co.aikar.commands.InvalidCommandArgument;
 import co.aikar.commands.PaperCommandManager;
 import com.archyx.aureliumskills.ability.AbilityManager;
 import com.archyx.aureliumskills.api.AureliumAPI;
-import com.archyx.aureliumskills.commands.ManaCommand;
-import com.archyx.aureliumskills.commands.SkillCommands;
-import com.archyx.aureliumskills.commands.SkillsCommand;
-import com.archyx.aureliumskills.commands.StatsCommand;
+import com.archyx.aureliumskills.commands.*;
 import com.archyx.aureliumskills.configuration.Option;
 import com.archyx.aureliumskills.configuration.OptionL;
 import com.archyx.aureliumskills.data.PlayerData;
@@ -61,9 +58,9 @@ import com.archyx.aureliumskills.skills.endurance.EnduranceLeveler;
 import com.archyx.aureliumskills.skills.excavation.ExcavationLeveler;
 import com.archyx.aureliumskills.skills.excavation.ExcavationLootHandler;
 import com.archyx.aureliumskills.skills.farming.FarmingAbilities;
+import com.archyx.aureliumskills.skills.farming.FarmingHarvestLeveler;
 import com.archyx.aureliumskills.skills.farming.FarmingInteractLeveler;
 import com.archyx.aureliumskills.skills.farming.FarmingLeveler;
-import com.archyx.aureliumskills.skills.farming.FarmingHarvestLeveler;
 import com.archyx.aureliumskills.skills.fighting.FightingAbilities;
 import com.archyx.aureliumskills.skills.fighting.FightingLeveler;
 import com.archyx.aureliumskills.skills.fishing.FishingAbilities;
@@ -515,7 +512,15 @@ public class AureliumSkills extends JavaPlugin {
 			}
 			return typeNames;
 		});
-		commandManager.registerCommand(new SkillsCommand(this));
+		commandManager.getCommandReplacements().addReplacement("skills_alias", "skills|sk|skill");
+		commandManager.registerCommand(new SkillsRootCommand(this));
+		commandManager.registerCommand(new ArmorCommand(this));
+		commandManager.registerCommand(new BackupCommand(this));
+		commandManager.registerCommand(new ItemCommand(this));
+		commandManager.registerCommand(new ModifierCommand(this));
+		commandManager.registerCommand(new ProfileCommand(this));
+		commandManager.registerCommand(new SkillCommand(this));
+		commandManager.registerCommand(new XpCommand(this));
 		commandManager.registerCommand(new StatsCommand(this));
 		commandManager.registerCommand(new ManaCommand(this));
 		if (OptionL.getBoolean(Option.ENABLE_SKILL_COMMANDS)) {
